Refactor ListActivities to new handler pattern
